[*] Formatted documentation comments according to doxygen style. Finished.

+++ Sort/Sort_Impl.hpp
@@ -660,7 +660,7 @@ void InsertionBinarySort<T>::sort(std::vector<T> & elements)
 {
      // TODO: implement InsertionBinary sort
     size_t length = elements.size();
-    for (int i = 1; i < length; i++)
+    for (size_t i = 1; i < length; i++)
     {
         T key = elements[i];
         int j = i - 1;
